+++ CHANGELOG.md
@@ -5,8 +5,12 @@
 Bug Fixes:
 
 1. Fix an exception when trying to display the MarkLogic rewriter endpoint line marker.
+2. Fix an exception in the roxy project configuration when there aren't any associated property files.
+
+Gradle Project Configuration:
+
+1. Support reading properties from `gradle.properties`.
 2. Fix an exception in the gradle project configuration when there aren't any associated property files.
-3. Fix an exception in the roxy project configuration when there aren't any associated property files.
 
 ## 2021